backend/requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
six==1.17.0
spotipy==2.25.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
